requirements.txt
alabaster==0.7.12
alembic==1.0.2
alipay==0.7.4
anaconda-client==1.7.2
anaconda-navigator==1.9.6
anaconda-project==0.8.2
aniso8601==4.0.1
appdirs==1.4.3
appnope==0.1.0
appscript==1.0.1
asn1crypto==0.24.0
astroid==2.1.0
atomicwrites==1.2.1
attrs==18.2.0
Automat==0.7.0
Babel==2.6.0
backcall==0.1.0
backports.os==0.1.1
backports.shutil-get-terminal-size==1.0.0
beautifulsoup4==4.6.3
bitarray==0.8.3
bleach==3.0.2
blinker==1.4
boto==2.49.0
certifi==2018.11.29
cffi==1.11.5
chardet==3.0.4
Click==7.0
cloudpickle==0.6.1
clyent==1.2.2
colorama==0.4.1
conda==4.5.12
conda-build==3.17.6
constantly==15.1.0
contextlib2==0.5.5
cryptography==2.3.1
cssselect==1.0.3
cycler==0.10.0
Cython==0.29.2
cytoolz==0.9.0.1
dask==1.0.0
decorator==4.3.0
defusedxml==0.5.0
distributed==1.25.1
docutils==0.14
dominate==2.3.4
entrypoints==0.2.3
et-xmlfile==1.0.1
fastcache==1.0.2
filelock==3.0.10
Flask==1.0.2
flask-blueprint==1.2.6
Flask-Bootstrap==3.3.7.1
Flask-Cache==0.13.1
Flask-Cors==3.0.7
Flask-DebugToolbar==0.10.1
Flask-Mail==0.9.1
flask-marshmallow==0.9.0
Flask-Migrate==2.3.0
Flask-RESTful==0.3.6
Flask-Script==2.0.6
Flask-Session==0.3.1
Flask-SQLAlchemy==2.3.2
gevent==1.3.7
glob2==0.6
gmpy2==2.0.8
greenlet==0.4.15
heapdict==1.0.0
html5lib==1.0.1
hyperlink==18.0.0
idna==2.7
imagesize==1.1.0
importlib-metadata==0.6
incremental==17.5.0
ipykernel==5.1.0
ipython==7.2.0
ipython-genutils==0.2.0
ipywidgets==7.4.2
isort==4.3.4
itsdangerous==1.1.0
jdcal==1.4
jedi==0.13.2
Jinja2==2.10
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.4
jupyter-console==6.0.0
jupyter-core==4.4.0
jupyterlab==0.35.3
jupyterlab-launcher==0.13.1
jupyterlab-server==0.2.0
keyring==17.0.0
kiwisolver==1.0.1
lazy-object-proxy==1.3.1
libarchive-c==2.8
lief==0.9.0
llvmlite==0.27.0
locket==0.2.0
lxml==4.2.5
Mako==1.0.7
MarkupSafe==1.1.0
marshmallow==2.16.3
mccabe==0.6.1
mistune==0.8.4
more-itertools==4.3.0
mpmath==1.1.0
msgpack==0.5.6
multipledispatch==0.6.0
navigator-updater==0.2.1
nbconvert==5.4.0
nbformat==4.4.0
networkx==2.2
nltk==3.4
nose==1.3.7
notebook==5.7.4
numpydoc==0.8.0
olefile==0.46
openpyxl==2.5.12
packaging==18.0
pandocfilters==1.4.2
parsel==1.5.1
parso==0.3.1
partd==0.3.9
path.py==11.5.0
pathlib2==2.3.3
pep8==1.7.1
pexpect==4.6.0
pickleshare==0.7.5
Pillow==5.3.0
pkginfo==1.4.2
pluggy==0.8.0
ply==3.11
prometheus-client==0.5.0
prompt-toolkit==2.0.7




psutil==5.4.8
ptyprocess==0.6.0
py==1.7.0
pyasn1==0.4.4
pyasn1-modules==0.2.2
pycodestyle==2.4.0
pycosat==0.6.3
pycparser==2.19
pycrypto==2.6.1
pycryptodome==3.7.0
pycryptodomex==3.7.0
pycurl==7.43.0.2
pyflakes==2.0.0
Pygments==2.3.1
PyHamcrest==1.9.0
pylint==2.2.2
PyMySQL==0.9.2
pyodbc==4.0.25
pyOpenSSL==18.0.0
pyparsing==2.3.0
PySocks==1.6.8
pytest==4.0.2
pytest-openfiles==0.3.1
pytest-remotedata==0.3.1
python-alipay-sdk==1.8.0
python-dateutil==2.7.5
python-editor==1.0.3
pytz==2018.7
PyYAML==3.13
pyzmq==17.1.2
qiniu==7.2.2
QtAwesome==0.5.3
qtconsole==4.4.3
QtPy==1.5.2
queuelib==1.5.0
redis==2.10.6
requests==2.20.1
rope==0.11.0
rsa==4.0
ruamel-yaml==0.15.46
schema==0.6.8
Send2Trash==1.5.0
service-identity==17.0.0
simplegeneric==0.8.1
simplejson==3.16.0
singledispatch==3.4.0.3
six==1.11.0
snowballstemmer==1.2.1
sortedcollections==1.0.1
sortedcontainers==2.1.0
Sphinx==1.8.2
sphinxcontrib-websupport==1.1.0
spyder==3.3.2
spyder-kernels==0.3.0
SQLAlchemy==1.2.14
sympy==1.3
tblib==1.3.2
terminado==0.8.1
testpath==0.4.2
toolz==0.9.0
tornado==5.1.1
tqdm==4.28.1
traitlets==4.3.2
Twisted==18.9.0
typed-ast==1.1.0
unicodecsv==0.14.1
urllib3==1.24.1
visitor==0.1.3
w3lib==1.19.0
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.14.1
widgetsnbextension==3.4.2
wincertstore==0.2
wrapt==1.10.11
wurlitzer==1.0.2
xlrd==1.1.0
XlsxWriter==1.1.2
xlwings==0.15.1
xlwt==1.3.0
xmltodict==0.11.0
zict==0.1.3
zope.interface==4.6.0
